requirements.txt
# Core NLP & Machine Learning
spacy>=3.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
nltk>=3.8.0

# Vector Search & Indexing
faiss-cpu>=1.7.0
# For GPU acceleration, use: faiss-gpu>=1.7.0

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0

# Text Processing & Cleaning
wordsegment>=1.3.0
emoji>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# LangChain & AI Integration
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Database Connectivity
pymysql>=1.1.0
sqlalchemy>=2.0.0

# OpenAI Integration
tiktoken>=0.5.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
